🔨 vscode: Change markdown ruler to 80
On my native mac screen, I have exactly 85 characters available per pane
when two are open in vertical split. Thus, change this back to 80, so we
don't have to scroll horizontally all the time to get the remaining 5
characters.

For Python (where I'm also using 90) that's not a big problem, as only a
few lines every reach the limit, but for plain text in markdown, this
occurs much more often.