zsh: Remove the git alias
This breaks running python in git, e.g. during an interactive rebase by
using
    $ rebase -i master --exec "tox -e linters"

as the locale setting is not fully qualified (the charset is missing).

I once had to overwrite the locale for the git command as it was always
showing the german command output not matter which locale was defined on
the system. By now, it looks like it's not needed anymore, so we are
just fine removing this.