Remove unused setting files for iTerm and PyCharm
The iTerm config changes with every update and I haven't updated the
file for a very long time. It's easier to just configure the necessary
settings as described in the `README.md` file when bootstrapping a new
machine.

As I haven't used PyCharm in ages, remove the settings as well. They are
most probably not compatible anymore anyway and since it's a `.jar`
archive the actual changes are hard to see as there isn't any git diff
available.